+++ pylinkage/optimization/__init__.py
@@ -1,3 +1,4 @@
 """Optimization package."""
 from .utils import generate_bounds
 from .grid_search import trials_and_errors_optimization
+from .particle_swarm import particle_swarm_optimization